Fixed spinning on enemy spinner graphic component + first attempt at rotation of lurker + changed delta time to double for better precision

+++ src/main/java/ryleh/view/PlayerGraphicComponent.java
@@ -105,7 +105,7 @@ private Direction direction;
 	}
 
 	@Override
-	public void render(final Point2D position, final int deltaTime) {
+	public void render(final Point2D position, final double deltaTime) {
 		rectangle.setX(position.getX());
 		rectangle.setY(position.getY());
 		this.updateImage(direction);
